# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Shell-Tunnel API",
    "description": "Ultra-lightweight remote shell gateway. Provides a cross-platform REST/WebSocket API for programmatic interaction with system shells.",
    "version": "0.7.0",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "iyulab",
      "url": "https://github.com/iyulab/shell-tunnel"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/v1/sessions": {
      "get": {
        "tags": [
          "Sessions"
        ],
        "summary": "List all sessions",
        "description": "Returns a list of all active sessions.\n\nRequired capability: `session.read`.",
        "operationId": "listSessions",
        "responses": {
          "200": {
            "description": "List of sessions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListSessionsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      },
      "post": {
        "tags": [
          "Sessions"
        ],
        "summary": "Create a new session",
        "description": "Creates a new shell session with optional configuration.\n\nRequired capability: `session.manage`.",
        "operationId": "createSession",
        "requestBody": {
          "description": "Session configuration",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Session created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
    "/api/v1/sessions/{sessionId}": {
      "get": {
        "tags": [
          "Sessions"
        ],
        "summary": "Get session status",
        "description": "Returns the current status of a session.\n\nRequired capability: `session.read`.",
        "operationId": "getSession",
        "parameters": [
          {
            "$ref": "#/components/parameters/sessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "Session status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      },
      "delete": {
        "tags": [
          "Sessions"
        ],
        "summary": "Delete a session",
        "description": "Terminates and removes a session.\n\nRequired capability: `session.manage`.",
        "operationId": "deleteSession",
        "parameters": [
          {
            "$ref": "#/components/parameters/sessionId"
          }
        ],
        "responses": {
          "204": {
            "description": "Session deleted successfully"
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
    "/api/v1/sessions/{sessionId}/execute": {
      "post": {
        "tags": [
          "Execution"
        ],
        "summary": "Execute command in session",
        "description": "Executes a command in an existing session.\n\nRequired capability: `exec`.",
        "operationId": "executeInSession",
        "parameters": [
          {
            "$ref": "#/components/parameters/sessionId"
          }
        ],
        "requestBody": {
          "description": "Command to execute",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteCommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteCommandResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid command or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
    "/api/v1/execute": {
      "post": {
        "tags": [
          "Execution"
        ],
        "summary": "Execute command (one-shot)",
        "description": "Executes a command without creating a persistent session.\n\nRequired capability: `exec`.",
        "operationId": "executeOneshot",
        "requestBody": {
          "description": "Command to execute",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteCommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteCommandResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid command or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
    "/api/v1/sessions/{sessionId}/ws": {
      "get": {
        "tags": [
          "Execution"
        ],
        "summary": "WebSocket connection for session",
        "description": "Establishes a WebSocket connection for real-time command streaming in a session.\n\nRequired capability: `exec`.",
        "operationId": "sessionWebSocket",
        "parameters": [
          {
            "$ref": "#/components/parameters/sessionId"
          }
        ],
        "responses": {
          "101": {
            "description": "WebSocket connection established"
          },
          "404": {
            "description": "Session not found"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
    "/api/v1/ws": {
      "get": {
        "tags": [
          "Execution"
        ],
        "summary": "WebSocket connection (one-shot)",
        "description": "Establishes a WebSocket connection for one-shot command execution.\n\nRequired capability: `exec`.",
        "operationId": "oneshotWebSocket",
        "responses": {
          "101": {
            "description": "WebSocket connection established"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque bearer token via the Authorization header. Each token carries a set of capabilities (exec, session.read, session.manage; '*' wildcard). A route declares a required capability: a missing/invalid token yields 401, a valid token lacking the capability yields 403. Role presets: operator, read-only, full-control."
      }
    },
    "parameters": {
      "sessionId": {
        "name": "sessionId",
        "in": "path",
        "required": true,
        "description": "Session identifier",
        "schema": {
          "type": "string"
        }
      }
    },
    "schemas": {
...
      "CreateSessionRequest": {
        "type": "object",
        "properties": {
          "shell": {
            "type": "string",
            "description": "Shell command to use (e.g., bash, powershell.exe)",
            "example": "bash"
          },
          "working_dir": {
            "type": "string",
            "description": "Initial working directory",
            "example": "/home/user"
          },
          "env": {
            "type": "object",
            "description": "Environment variables to set",
            "additionalProperties": {
              "type": "string"
            },
            "example": {
              "MY_VAR": "my_value"
            }
          }
        }
      },
      "CreateSessionResponse": {
        "type": "object",
        "required": [
          "session_id",
          "session_id_str"
        ],
        "properties": {
          "session_id": {
            "type": "integer",
            "format": "int64",
            "description": "Numeric session ID",
            "example": 1
          },
          "session_id_str": {
            "type": "string",
            "description": "Human-readable session ID",
            "example": "session-1"
          }
        }
      },
...
      "ListSessionsResponse": {
        "type": "object",
        "required": [
          "count",
          "sessions"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "description": "Number of sessions",
            "example": 2
          },
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionSummary"
            }
          }
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication required: the Authorization: Bearer token is missing or invalid (empty body)."
      },
      "ForbiddenError": {
        "description": "The bearer token is valid but lacks the capability this route requires (empty body)."
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}